backend/requirements.txt
# PhotoRank AI — Python Backend Dependencies
# Install: pip install -r requirements.txt

fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
pillow>=10.0.0
numpy>=1.24.0
rawpy>=0.19.0
pytest==8.3.0
httpx==0.27.0
pytest-asyncio==0.24.0
